requirements.txt
discord.py[voice]
Pillow
flask
requests
aiohttp
jsons
yt-dlp[default]